unit-test fixes
