+++ fcs_aux/tests/unit/test_resident_acting_flow.py
@@ -48,10 +48,11 @@ def test_mark_as_executing():
         update_processing_mock.assert_called_once_with(True)
 
 
-def test_start_action():
+def test_start_processing():
     resident_id = "3j"
-    resident = Resident(dict(_id=resident_id))
+    resident = Resident(dict(_id=resident_id, energy=1000))
     bldg = {
+        "flr": "g-b(1,2)-l0",
         "payload": {
             "da": "nyet"
         }
@@ -59,7 +60,7 @@ def test_start_action():
     action = "fetch-article"
     with patch("mies.residents.acting.flow.add_new_action_status") as add_status_mock:
         with patch("mies.residents.acting.flow.app.send_task") as send_task_mock:
-            resident.start_action(action, bldg)
+            resident.start_processing(action, bldg)
             send_task_mock.assert_called_once_with(action, [bldg["payload"]])
             add_status_mock.assert_called_once_with(bldg, ANY)
             action_status = add_status_mock.call_args[0][1]
